Reorganize function and developement of command lines tools for skekt
