+++ setup.py
@@ -87,7 +87,7 @@ setup(
     # For a discussion on single-sourcing the version across setup.py and the
     # project code, see
     # https://packaging.python.org/en/latest/single_source_version.html
-    version='0.0.1',  # Required
+    version='0.0.2',  # Required
 
     # This is a one-line description or tagline of what your project does. This
     # corresponds to the "Summary" metadata field:
@@ -184,7 +184,7 @@ setup(
 
     # If there are data files included in your packages that need to be
     # installed, specify them here.
-    package_data={'wrfvis': ['data/template.html',
+    package_data={'wrfvis': ['data/template.html','data/template_skewt.html','data/template_skewt_delta.html'
                               ]
                   },
 
@@ -203,7 +203,7 @@ setup(
     # executes the function `main` from this package when invoked:
     entry_points={  # Optional
         'console_scripts': [
-            'wrfvis_gridcell=wrfvis.cltools:wrfvis_gridcell',
+            'wrfvis_gridcell=wrfvis.cltools:wrfvis_gridcell',  'wrfvis_skewt=wrfvis.cltools:wrfvis_skewt',
         ],
     },
