+++ wrfvis.egg-info/PKG-INFO
@@ -1,6 +1,6 @@
 Metadata-Version: 2.1
 Name: wrfvis
-Version: 0.0.1
+Version: 0.0.2
 Summary: A WRF model output visualization software
 Home-page: 
 Author: Manuela Lehner
