+++ wrfvis.egg-info/SOURCES.txt
@@ -15,6 +15,8 @@ wrfvis.egg-info/entry_points.txt
 wrfvis.egg-info/requires.txt
 wrfvis.egg-info/top_level.txt
 wrfvis/data/template.html
+wrfvis/data/template_skewt.html
+wrfvis/data/template_skewt_delta.html
 wrfvis/tests/__init__.py
 wrfvis/tests/test_cltools.py
 wrfvis/tests/test_core.py
