+++ wrfvis/core.py
@@ -8,7 +8,7 @@ import numpy as np
 import pandas as pd
 import xarray as xr
 
-from wrfvis import cfg, grid, graphics
+from wrfvis import cfg, grid, graphics, skewT
 
 
 def get_wrf_timeseries(param, lon, lat, zagl=None):
@@ -168,6 +168,208 @@ def write_html(param, lon, lat, zagl, directory=None):
                 txt = txt.replace('[PLOTVAR]', param)
                 txt = txt.replace('[IMGTYPE]', 'timeseries')
                 out.append(txt)
+            with open(outpath, 'w') as outfile:
+                outfile.writelines(out)
+
+        return outpath
+
+
+def write_html_skewt(time, lon, lat, directory=None):
+    '''
+    Create an html file to plot a single Skew T-logP plot with wind profile,
+    hodographs and Skew T-logP indices.
+
+    Author
+    ------
+    Christian Brida
+
+    Parameters
+    ----------
+    time : str
+        timestamp, use the format YYYY-MM-DDTHH:MM.
+    lon : float
+        the longitude
+    lat : float
+        the latitude
+    directory : str, optional
+        directory where the html file is saved. The default is None.
+
+    Returns
+    -------
+    outpath : str
+        filepath.
+
+    '''
+
+    if os.path.exists(cfg.wrfout):
+        # create directory for the plot
+        if directory is None:
+            directory = mkdtemp()
+        mkdir(directory)
+
+        print('Plotting topography')
+        hgt = skewT.get_hgt(lon, lat)
+        topo = os.path.join(directory, 'topo.png')
+        graphics.plot_topo(hgt, (lon, lat), filepath=topo)
+
+        print('Plotting Skew T-log P')
+        skewt = os.path.join(directory, 'skewt.png')
+        graphics.plot_skewt(time, lon, lat, filepath=skewt)
+
+        print('Plotting wind profile')
+        wind = os.path.join(directory, 'wind.png')
+        graphics.plot_wind_profile(time, lon, lat, filepath=wind)
+
+        print('Plotting hodograph')
+        hodo = os.path.join(directory, 'hodo.png')
+        graphics.plot_hodograph(time, lon, lat, filepath=hodo)
+
+        # plot_skewt_full(time, lon, lat, filepath=skewt)
+
+        print('Parameters')
+
+        FREEZING_LEVEL_m, PRECIP_WATER, TOTAL_TOTALS_INDEX, RH_0 = skewT.calc_skewt_param_general(
+            time, lon, lat)
+
+        ML_LCL, ML_LFC, ML_LI, ML_CAPE, ML_CIN = skewT.calc_skewt_param_mixed_layer(
+            time, lon, lat)
+
+        SB_LCL, SB_LFC, SB_LI, SB_CAPE, SB_CIN = skewT.calc_skewt_param_surface_based(
+            time, lon, lat)
+
+        RM_DIR, RM_SPEED, SHEAR_1KM, SHEAR_6KM, SRH_1km_tot, SRH_3km_tot = skewT.calc_skewt_param_wind(
+            time, lon, lat)
+
+        MUCAPE, EL, CAPE_strenght, K_INDEX = skewT.calc_skewt_param_extra(
+            time, lon, lat)
+        # create HTML from template
+        outpath = os.path.join(directory, 'index.html')
+        with open(cfg.html_template_skewt, 'r') as infile:
+            lines = infile.readlines()
+            out = []
+            for txt in lines:
+                ''' Coordinates'''
+                txt = txt.replace('[LAT]',
+                                  f'{lat}')
+                txt = txt.replace('[LON]',
+                                  f'{lon}')
+                txt = txt.replace('[TIME]',
+                                  f'{time}')
+
+                ''' General parameters '''
+                txt = txt.replace('[FREEZING_LEVEL_m]',
+                                  f'{FREEZING_LEVEL_m:.0f}')
+                txt = txt.replace('[PRECIP_WATER]',
+                                  f'{PRECIP_WATER.magnitude:.2f}')
+                txt = txt.replace('[TOTAL_TOTALS_INDEX]',
+                                  f'{TOTAL_TOTALS_INDEX.magnitude:.2f}')
+                txt = txt.replace('[RH_0]', f'{RH_0.magnitude:.2f}')
+
+                ''' Mixed Layer parcel '''
+                txt = txt.replace('[ML_LCL]', f'{ML_LCL.magnitude:.2f}')
+                txt = txt.replace('[ML_LFC]', f'{ML_LFC.magnitude:.2f}')
+                txt = txt.replace('[ML_LI]', f'{ML_LI[0].magnitude:.2f}')
+                txt = txt.replace('[ML_CAPE]', f'{ML_CAPE.magnitude:.2f}')
+                txt = txt.replace('[ML_CIN]', f'{ML_CIN.magnitude:.2f}')
+
+                ''' Surface based parcel '''
+                txt = txt.replace('[SB_LCL]', f'{SB_LCL.magnitude:.2f}')
+                txt = txt.replace('[SB_LFC]', f'{SB_LFC.magnitude:.2f}')
+                txt = txt.replace('[SB_LI]', f'{SB_LI[0].magnitude:.2f}')
+                txt = txt.replace('[SB_CAPE]', f'{SB_CAPE.magnitude:.2f}')
+                txt = txt.replace('[SB_CIN]', f'{SB_CIN.magnitude:.2f}')
+
+                ''' Wind indices '''
+                txt = txt.replace('[RM_DIR]', f'{RM_DIR:.2f}')
+                txt = txt.replace('[RM_SPEED]', f'{RM_SPEED:.2f}')
+                txt = txt.replace('[SHEAR_1KM]', f'{SHEAR_1KM:.2f}')
+                txt = txt.replace('[SHEAR_6KM]', f'{SHEAR_6KM:.2f}')
+                txt = txt.replace(
+                    '[SRH_1km_tot]', f'{SRH_1km_tot.magnitude:.2f}')
+                txt = txt.replace(
+                    '[SRH_3km_tot]', f'{SRH_3km_tot.magnitude:.2f}')
+
+                ''' Other indices '''
+                txt = txt.replace('[MUCAPE]', f'{MUCAPE.magnitude:.2f}')
+                txt = txt.replace('[EL]', f'{EL.magnitude:.2f}')
+                txt = txt.replace('[CAPE_strenght]',
+                                  f'{CAPE_strenght.magnitude:.2f}')
+                txt = txt.replace('[K_INDEX]', f'{K_INDEX.magnitude:.2f}')
+
+                out.append(txt)
+            with open(outpath, 'w') as outfile:
+                outfile.writelines(out)
+
+        return outpath
+
+
+def write_html_delta_skewt(time, lon, lat, deltatime, directory=None):
+    '''
+    Create an html file to plot a delta Skew T-logP plot with wind profile,
+    hodographs and Skew T-logP indices.
+
+    Author
+    ------
+    Christian Brida
+
+    Parameters
+    ----------
+    time : str
+        timestamp, use the format YYYY-MM-DDTHH:MM.
+    lon : float
+        the longitude
+    lat : float
+        the latitude
+    deltatime : int, optional
+        delta time in hours from time. The default is 24. units: h.
+    directory : str, optional
+        directory where the html file is saved. The default is None.
+
+    Returns
+    -------
+    outpath : str
+        filepath.
+
+    '''
+    if os.path.exists(cfg.wrfout):
+        # create directory for the plot
+        if directory is None:
+            directory = mkdtemp()
+            mkdir(directory)
+
+            print('Plotting topography')
+            hgt = skewT.get_hgt(lon, lat)
+            topo = os.path.join(directory, 'topo.png')
+            graphics.plot_topo(hgt, (lon, lat), filepath=topo)
+
+            print('Plotting Skew T-log P delta')
+            # plot the timeseries
+            skewt_delta = os.path.join(directory, 'skewt_delta.png')
+            graphics.plot_skewt_deltatime(time, lat, lon, deltatime,
+                                          filepath=skewt_delta)
+
+            print('Plotting Skew T-log P avg')
+            # plot the timeseries
+            skewt_avg = os.path.join(directory, 'skewt_avg.png')
+            graphics.plot_skewt_averaged(
+                time, lat, lon, deltatime, filepath=skewt_avg)
+
+            # create HTML from template
+            outpath = os.path.join(directory, 'index.html')
+            with open(cfg.html_template_skewt_delta, 'r') as infile:
+                lines = infile.readlines()
+                out = []
+                for txt in lines:
+                    txt = txt.replace('[LAT]',
+                                      f'{lat}')
+                    txt = txt.replace('[LON]',
+                                      f'{lon}')
+                    txt = txt.replace('[TIME]',
+                                      f'{time}')
+                    txt = txt.replace('[DELTATIME]',
+                                      f'{deltatime}')
+
+                    out.append(txt)
             with open(outpath, 'w') as outfile:
                 outfile.writelines(out)
